+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Added .gitignore to prevent generated files from being committed.
 - Added build.bat to automatically build an executable using PyInstaller.
+- Added Verify.py to help verify the upstream text files and find any entries that aren't formatted correctly.
 
 ### Changes